UserService/requirements.txt
fastapi==0.111.0
email_validator==2.1.1
fastapi-cli==0.0.4
httpx==0.25.2
Jinja2==3.1.4
orjson==3.10.5
pydantic==2.7.4
python-multipart==0.0.9
starlette==0.37.2
typing_extensions==4.12.2
ujson==5.10.0
uvicorn==0.30.1
httptools==0.6.1
pipdeptree==2.22.0
python-dotenv==1.0.1
PyYAML==6.0.1
watchfiles==0.22.0
websockets==12.0
passlib==1.7.4
bcrypt==4.1.3
PyJWT==2.8.0
pyuploadcare==5.1.0
python-dotenv==1.0.1
pymongo==4.7.3